figures out border of polyminoe
I've decided to implement the algorithm Dhruv and I were discussing last friday. Will create visualization of the piece, as well as the outer perimeter, through a 2D array of chars. I've decided to use chars and not integers for reasons I can discuss more thoroughly Friday. Currently it does not count the number of outer permeter pieces but the code can be easily modified to make that change.